API for meta string
